Remove std::move calls from SyncRoutes80211s
Summary: A `std::move` call tells the compiler that we're never going to use the variable again, but in this case we do end up continuing to use those variables and even call `std::move` again. This ends up being undefined behavior that has just worked up until now, but we don't have that guarantee so let's fix it.

Reviewed By: ilyacodesFB

Differential Revision: D17789234

fbshipit-source-id: 30837ca7578365c745fccbe05361c8386b052536