Stop GatewayConnectivityMonitor from blocking main thread
Summary:
Previous changes made to `GatewayConnectivityMonitor` added it to the same thread as other classes. But, because the `GCM` contains a blocking socket call, this blocks the entire thread.

This diff fixes that issue by giving it its own thread. In the future, it will be made non blocking and moved back to the main thread.

Reviewed By: ilyacodesFB

Differential Revision: D17835833

fbshipit-source-id: 9fe90aeadfac759d5bddd44e908007615152765f