Restructure SyncRoutes80211s to use folly::EventBase
Summary: Same old, same old, moving another class away from `fbzmq` over to `folly`. The slight issue with this one is that it does require the `rnl` library, which has not been moved over, so there's a bit of restructuring to build that in the main function to at least consolidate `fbzmq`.

Reviewed By: ilyacodesFB

Differential Revision: D17456121

fbshipit-source-id: 67e9dacf48fe5492a13270aae34c9a538dbd8aee