Add support for debugfs-like stats export
Summary: It can be useful for manual debugging to see the state of `fbmeshd`. In order to facilitate this, add a new option that allows stats to be written to a user-specified folder. This is similar in style to the Linux concept of debugfs, except that the files actually exist.

Reviewed By: mariof

Differential Revision: D18089762

fbshipit-source-id: 4729ef86066266996e2b18391ddc677e045aed4c